Second revisions suggested by Krista
